+++ src/main/java/project/backend/matching/MatchingController.java
@@ -2,6 +2,7 @@ package project.backend.matching;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
+import project.backend.matching.dto.MatchingResultDTO;
 
 @RestController
 @RequestMapping("/matching")
@@ -11,7 +12,7 @@ public class MatchingController {
     private final MatchingService matchingService;
 
     @GetMapping("/{userId}")
-    public  MatchingResultDTO getMatchingResult(@PathVariable("userId") Long userId) throws Exception {
+    public MatchingResultDTO getMatchingResult(@PathVariable("userId") Long userId) throws Exception {
         return matchingService.getMatchingResult(userId);
     }
 }
