+++ src/main/java/project/backend/matching/MatchingService.java
@@ -3,11 +3,12 @@ package project.backend.matching;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import project.backend.matching.dto.MatchingResultDTO;
 import project.backend.mypage.dto.MyPageDisplayDTO;
 import project.backend.openai.OpenAiService;
-import project.backend.pythonapi.PersonInfo;
-import project.backend.pythonapi.SajuRequest;
-import project.backend.pythonapi.SajuResponse;
+import project.backend.pythonapi.dto.PersonInfo;
+import project.backend.pythonapi.dto.SajuRequest;
+import project.backend.pythonapi.dto.SajuResponse;
 import project.backend.pythonapi.SajuService;
 import project.backend.user.UserRepository;
 import project.backend.user.dto.UserEnums;
