+++ src/main/java/project/backend/openai/OpenAiController.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import project.backend.openai.dto.PromptRequest;
 
 @RestController
 @RequiredArgsConstructor
