+++ src/main/java/project/backend/openai/dto/PromptRequest.java
@@ -1,4 +1,4 @@
-package project.backend.openai;
+package project.backend.openai.dto;
 
 public record PromptRequest(String prompt) {
 }
