+++ src/main/java/project/backend/pythonapi/SajuController.java
@@ -5,6 +5,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import project.backend.pythonapi.dto.SajuRequest;
+import project.backend.pythonapi.dto.SajuResponse;
 import reactor.core.publisher.Mono;
 
 @RestController
