+++ src/main/java/project/backend/pythonapi/SajuService.java
@@ -3,6 +3,8 @@ package project.backend.pythonapi;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
+import project.backend.pythonapi.dto.SajuRequest;
+import project.backend.pythonapi.dto.SajuResponse;
 import reactor.core.publisher.Mono;
 
 @Service
