+++ src/main/java/project/backend/pythonapi/dto/PersonInfo.java
@@ -1,4 +1,4 @@
-package project.backend.pythonapi;
+package project.backend.pythonapi.dto;
 
 public record PersonInfo(int year, int month, int day, int gender) {
